package/mender: add BR2_PACKAGE_MENDER_ARTIFACT_NAME option
ARTIFACT_NAME is used to set a unique string of the built artifact/image
and this will be reported to the Mender managment server as "current
software running".

The Mender managment server will skip updating device if the
ARTIFACT_NAME is the same on the target device as in the Mender artifact
and hence this name must be unique for every release.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>