package/mender: add BR2_PACKAGE_MENDER_DEVICE_TYPE option
MENDER_DEVICE_TYPE is used to verify compatibility of Mender artifacts
against devices and it will also be reported to the Mender management
server from the device for the server to be able to verify this.

Device type can not change once a device has been provisioned with the
initial software and this is why it is stored on the persistent data
partition which is preserved across updates.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>