package/mender: add BR2_PACKAGE_MENDER_UPDATE_MODULE option
Update Modules are small extensions to the Mender client
which allows definition of custom install/update types.

This feature was introduced in Mender client 2.0.0, and
there are a couple of reference scripts provided in the
source tree which can be installed to test this feature.

The reference Update Modules are bash script, and to
avoid having a hard dependecy on 'bash' an build time
option is added to install the reference Update Modules
if needed.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>